+++ lib/shrine/storage/s3.rb
@@ -427,33 +427,43 @@ class Shrine
 
       # Uploads the file to S3. Uses multipart upload for large files.
       def put(io, id, **options)
-        if io.respond_to?(:path)
-          path = io.path
-        elsif io.is_a?(UploadedFile) && defined?(Storage::FileSystem) && io.storage.is_a?(Storage::FileSystem)
-          path = io.storage.path(io.id).to_s
-        end
+        bytes_uploaded = nil
 
-        if path
+        if (path = extract_path(io))
           # use `upload_file` for files because it can do multipart upload
           options = { multipart_threshold: @multipart_threshold[:upload] }.merge!(options)
           object(id).upload_file(path, **options)
-          File.size(path)
+          bytes_uploaded = File.size(path)
         else
           io.to_io if io.is_a?(UploadedFile) # open if not already opened
 
-          if io.respond_to?(:size) && io.size
+          if io.respond_to?(:size) && io.size && (io.size <= @multipart_threshold[:upload] || !object(id).respond_to?(:upload_stream))
             object(id).put(body: io, **options)
-            io.size
+            bytes_uploaded = io.size
+          elsif object(id).respond_to?(:upload_stream)
+            # `upload_stream` uses multipart upload
+            object(id).upload_stream(tempfile: true, **options) do |write_stream|
+              bytes_uploaded = IO.copy_stream(io, write_stream)
+            end
           else
-            # IO has unknown size, so we have to use multipart upload
-            multipart_put(io, id, **options)
+            Shrine.deprecation "Uploading a file of unknown size with aws-sdk-s3 older than 1.14 is deprecated and will be removed in Shrine 3. Update to aws-sdk-s3 1.14 or higher."
+
+            Tempfile.create("shrine-s3", binmode: true) do |file|
+              bytes_uploaded = IO.copy_stream(io, file.path)
+              object(id).upload_file(file.path, **options)
+            end
           end
         end
+
+        bytes_uploaded
       end
 
-      # Uploads the file to S3 using multipart upload.
-      def multipart_put(io, id, **options)
-        MultipartUploader.new(object(id)).upload(io, **options)
+      def extract_path(io)
+        if io.respond_to?(:path)
+          io.path
+        elsif io.is_a?(UploadedFile) && defined?(Storage::FileSystem) && io.storage.is_a?(Storage::FileSystem)
+          io.storage.path(io.id).to_s
+        end
       end
 
       # The file is copyable if it's on S3 and on the same Amazon account.
@@ -478,52 +488,6 @@ class Shrine
       def encode_content_disposition(content_disposition)
         content_disposition.sub(/(?<=filename=").+(?=")/) do |filename|
           CGI.escape(filename).gsub("+", " ")
-        end
-      end
-
-      # Uploads IO objects of unknown size using the multipart API.
-      class MultipartUploader
-        def initialize(object)
-          @object = object
-        end
-
-        # Initiates multipart upload, uploads IO content into multiple parts,
-        # and completes the multipart upload. If an exception is raised, the
-        # multipart upload is automatically aborted.
-        def upload(io, **options)
-          multipart_upload = @object.initiate_multipart_upload(**options)
-
-          parts = upload_parts(multipart_upload, io)
-          bytes_uploaded = parts.inject(0) { |size, part| size + part.delete(:size) }
-
-          multipart_upload.complete(multipart_upload: { parts: parts })
-
-          bytes_uploaded
-        rescue
-          multipart_upload.abort if multipart_upload
-          raise
-        end
-
-        # Uploads parts until the IO object has reached EOF.
-        def upload_parts(multipart_upload, io)
-          1.step.inject([]) do |parts, part_number|
-            parts << upload_part(multipart_upload, io, part_number)
-            break parts if io.eof?
-            parts
-          end
-        end
-
-        # Uploads at most 5MB of IO content into a single multipart part.
-        def upload_part(multipart_upload, io, part_number)
-          Tempfile.create("shrine-s3-part-#{part_number}", binmode: true) do |body|
-            IO.copy_stream(io, body, MIN_PART_SIZE)
-            body.rewind
-
-            multipart_part = multipart_upload.part(part_number)
-            response       = multipart_part.upload(body: body)
-
-            { part_number: part_number, size: body.size, etag: response.etag }
-          end
         end
       end
     end